rename subpackage 'featureWriters' in plural
make liststr a staticmethod
