+++ Lib/ufo2ft/featureWriters/__init__.py
@@ -1,0 +1,7 @@
+from .baseFeatureWriter import BaseFeatureWriter
+from .kernFeatureWriter import KernFeatureWriter
+from .markFeatureWriter import MarkFeatureWriter
+
+
+def defaultFeatureWriters(font):
+    return [KernFeatureWriter(font), MarkFeatureWriter(font)]
